Give the binding tests a class of their own.
Tuple destructuring is not a TensorFlow operation, so the minimal pair sat in `TestDatasets` only because it arrived beside the corpus-shaped one, which does involve datasets and stays. `TestTupleBinding` holds the forms where the question is whether a type survives a binding rather than what an API computes.
